Fix AFA -> NFA conversion
The error was in converting the NNFA to an NFA,
and not in the AFA -> NNFA conversion.

Before, I naïvely just connected one of the start states `s ∈ S`
to all of the start states `S` of the NNFA with ε-transitions,
and made `s` the start state of the NFA.
The problem with this is that there might be connections *to* `s`,
and now we can get from `s` to any other state in `S` "for free".

The solution is to introduce a new dedicated start state `s_0`
and connect *that* to all states in *S* with ε-transitions.