CNF and DNF types
They are much like the `ListList` type, but with better type-checking
guarantees. I thought I could make a more "principled" Monad instance
for either, but I was unsuccessful, so we are using the same "hack" is
in the `ListList` instance where we get away with it by making the `Eq`
instance convert to sets before comparing, so order and duplicates are
disregarded (if they are not, then some of the monad laws are broken).

Also remove the `ListListAutomaton` type (and associated module),
as using either the `CNF` or `DNF` types will convey a better meaning.
Some would say the same should apply to the `ListAutomaton`, but that
will have to wait.