Fix bug in 2-stack PDA automaton implementation
When we found a ε-loop, we would only check for the states that were
actually part of the loop if it was final, but we should really be
checking every state since we last read a symbol. For example if,
after having read all input, took the following path with ε-transitions:

  A --> B --> C
        ^     |
        |     |
        +-----+

we should check if state A, B and C are final, not just state B and C.

Also rename the `beenHere` function to `dejavu`,
as we are not actually checking if we have been in the same
configuration before, but merely if we have been in a *similar*
confuguration.