Make 2sDPDA fully functional
The transition function, instead of returning `Maybe a` as a way to
potentially not consume the input, now returns a Boolean flag signaling
whether or not to consume the input.
This means that the transition function is not partial anymore.

Also limit the amount of symbols that could be pushed to the stack to <3
rather than the unlimited amount allowed before.