New idea for FPDA
After reading all input, we essentially run a DFA on the remaining stack
to decide if we accept or not.

Also change to use a type of PDA that *always* reads a stack symbol, and
then starts with an initial symbol on the stack.

The conversion from Sipser DPDA is completely untested and likely very
buggy.