Additional DPDA test
Tests for recognizing a string "mirrored" around a certain character.
Can also be described as a palindrome with a dedicated character in the
middle.

This has revealed a bug in the `toSipserDPDA` function.
The problem is that the γ function of the FPDA might move
over a final state, which should *not* accept if it ends up
in a non-final state.
But, the Sipser DPDA *will* accept on such a move.
The solution is to introduce a dedicated final state,
and a dedicated "bottom-of-stack" symbol, and only
move to this final state if we have read all input,
we have reached the bottom-of-stack symbol,
and we are currently in a final state of the FPDA.