[WIP] Use FPDA as base for Monadic PDA
There is a bug in the implementation of `fromSipserNPDA`,
and i still need to implement `toSipserNPDA` in the
`Automata.PushDown.Monadic.List` module.

I'm also debating whether or not to keep the `acceptsAngelic`
and `acceptsDemonic` functions as we implicity assume for the most part
that we are working with angelic non-determinism.
To be totally correct I should probably use a dedicated `Disjunction`
and `Conjunction` monad and keep each in their own modules.