Fix `Automata.PushDown.Monadic.List.fromSipserNPDA`
Before, the γ function would return states that didn't actually read the
top stack symbol, or alternatively wrote new symbols on top of the
stack, after reading the stack symbol.

In either case, these would not "uncover" the next stack symbol,
and so it is an error to include them in the result of γ.

Not we only return the states from configurations with an empty stack,
or if we find a final state we return that, leading to accepting the
string.