New definition of Functional PDA
Based in the read-pop PDA from section 15 of [1].

Still need to think about how to convert to and from Sipser DPDAs.
There are some parts that are going to be tricky, such as Sipser DPDAs
accepting if *any* state is reachable after reading all input,
and the fact that a Sipser DPDA can, without reading additional input,
pop an some amount from the stack and then push some amount to the
stack, whereas this FPDA can only push to the stack when reading input.

[1]: F. Henglein, “Notes on Deterministic Push-Down Automata.” Jul. 20, 2023.