Modify FPDA definition yet again
Now follows the definition from [1] even more closely.

The translation from this FPDA to Sipser DPDA seems fairly simple,
but the other direction is a little more tricky.
I think I'm on the right track though,
but as the tests show there is still parts that need to be fixed.

[1]: F. Henglein, “Notes on Deterministic Push-Down Automata.” Jul. 20, 2023.