Sipser DPDA to FPDA conversion.
It required a slight modification to the FPDA,
namely that δ_read can now transition to read states
and push to the stack while doing so.

Technically this change was not necessary, but it makes
the SDPDA -> FPDA function quite a bit simpler,
and it doesn't make the semantics of the FPDA much more complicated.