Re-work `fromSipserDPDA` and fix `fromSipserNPDA`
The function now only simulates the ε-transitions of the DPDA *after*
doing a transition on input, whereas before it did it both before and
after. This does necessitate a dedicated start state, as we do need to
simulate potential ε-transitions until we find a proper read state.

The `fromSipserNPDA` is awfully slow now, but I suspect that is
(at least in part) because it uses a slower version of
`SipserNPDA.stepE`, which keeps track of if any final states were
encountered along the way.