ListList finite-state automaton
Corresponds somewhat to an AFA, except that the AFA can also do
negations, so we need two times the states when converting from
an AFA to ListListFA, and so it lies somewhere between the ListFA and
PropositionFA in terms of state complexity.

It would be good to have some bespoke example test cases rather than
just random tests, but the size quickly blows up.