Probability monad FA and PDA
Add definitions for monadic FA and PDA with the Probability monad,
in effect implementing probabilistic FAs and PDAs.

Also add definition for monadic PDA with the Proposition monad
(equivalent to alternating PDAs) as it was missing.

Include test for probability FA that shows it can recognize the
non-regular stochastic language shown in:
https://www-verimag.imag.fr/~iosif/LogicAutomata07/CoursStochLang.pdf#Navigation12