Operations on List PDAs.
Can now take the union and inversion of List PDAs.

This is tested with the non-CFL {w·w | w ∈ Σ*},
whose complement *is* CFL.
The test shows that, with angelic non-determinism
a List PDA can recognize the complement of the language,
and when inverted with demonic non-determinism,
the List PDA can recognize the language.

The PDA that recognizes the complement also uses
the `union` function, so it is in a way also an indirect
test of that function.

It is then clear that we can also implement an `intersect`
function for List PDAs that operate with demonic non-determinism.