No empty probabilities allowed in PFAs and PPDAs
Remove transitions to "empty" probabilities, i.e. `fromFreqs []`
in example FA and PDA, and remove the check in the acceptance function.
While empty transitions works fine for non-deterministic FAs and PDAs to
signal rejection without introducing a dedicated rejection state, it does
not play nice with the probability monad.

Also improve the quality of generated test cases for the complement of the EQ
language.