Change ways ε-closure of PDAs are computed
This comes from a revelation that currently,
the way the PDA is implemented does not follow the conventional
semantics (i.e. PDA accepts iff. there exists a path of configurations
that end in a final state after reading all input),
as we cut ε-cycles short.
But cutting them short *does* change the semantics,
as demonstrated by the `npdaEpsilonCycle` example.
Therefore, we change the semantics of the PDA,
as implementing the conventional semantics is difficult.